requirements.txt
numpy==1.25.2
opencv-python==4.8.1.78
packaging==23.1
pillow==8.4.0